Issue #380

* Pull in: obstack, adt from newlib
* Replace: ctype, balloc with code from newlib
* Add compatibility API for: balloc
* Deprecated classic adt and balloc apis
* Implement asm_realloc, asm_free so that adt works
* Small rearrange of classic folders

+++ libsrc/fcntl/cpm/write.c
@@ -14,7 +14,7 @@
 #include <cpm.h>
 
 
-unsigned int write(int fd, void *buf, size_t len)
+ssize_t write(int fd, void *buf, size_t len)
 {   
     char    buffer[SECSIZE+2];
     unsigned char uid;
